LayoutTests/imported/w3c/web-platform-tests/fetch/api/basic/status.h2.any-expected.txt


FAIL statusText over H2 for status 200 should be the empty string assert_equals: statusText should be the empty string expected "" but got "OMG"
FAIL statusText over H2 for status 210 should be the empty string assert_equals: statusText should be the empty string expected "" but got "OMG"
FAIL statusText over H2 for status 400 should be the empty string assert_equals: statusText should be the empty string expected "" but got "OMG"
FAIL statusText over H2 for status 404 should be the empty string assert_equals: statusText should be the empty string expected "" but got "OMG"
FAIL statusText over H2 for status 410 should be the empty string assert_equals: statusText should be the empty string expected "" but got "OMG"
FAIL statusText over H2 for status 500 should be the empty string assert_equals: statusText should be the empty string expected "" but got "OMG"
FAIL statusText over H2 for status 502 should be the empty string assert_equals: statusText should be the empty string expected "" but got "OMG"
